refactor: convert registration handlers to use UnitOfWork pattern
